// frontend/lib/api/models/EOTask2Response.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Response sujet Tâche 2.
 */
export type EOTask2Response = {
    id: string;
    session_id: string;
    subject: string;
    order: number;
    created_at: string;
    updated_at: string;
};